added 2 redux files for project, app - removed actions from headerRedux, registered new reducers
