+++ apps/src/code-studio/appRedux.js
@@ -1,0 +1,31 @@
+const SET_APP_LOAD_STARTED = 'header/SET_APP_LOAD_STARTED';
+const SET_APP_LOADED = 'header/SET_APP_LOADED';
+
+const initialState = {
+  appLoading: false,
+  appLoaded: false
+};
+
+export default (state = initialState, action) => {
+  if (action.type === SET_APP_LOAD_STARTED) {
+    return {
+      ...state,
+      appLoadStarted: true
+    };
+  }
+  if (action.type === SET_APP_LOADED) {
+    return {
+      ...state,
+      appLoaded: true
+    };
+  }
+  return state;
+};
+
+export const setAppLoadStarted = () => ({
+  type: SET_APP_LOAD_STARTED
+});
+
+export const setAppLoaded = () => ({
+  type: SET_APP_LOADED
+});
